refactor: 게시판 관련 페이지 pages/board/ui 폴더로 이동
