+++ apps/web/src/pages/board/ui/AccompanyDetailPage.tsx
