+++ apps/web/src/pages/board/ui/CommunityDetailPage.tsx
